Updated Board class methods get/setValue to Integer objects to better interact with conversion of String types in JTextFields; Updated draw method in KataView to display boards intialized state at launch and reduced field length to two. Updated main to exit JFrame when window is closed. Updated the mouse clicked event to monitor for at least three cell's vital statuses to be alive before allowing a simulation to begin, and disable further changes by hiding fields upon criteria met or simulating variable no longer false.

+++ test/src/com/comresource/kata/TestView.java
@@ -31,11 +31,11 @@ public class TestView {
 	@Test
 	public void testKataContruct() {
 		assertEquals(false, k.simulating);		
-		assertEquals(3, k.lifeBoard.getCols());
-		assertEquals(3, k.lifeBoard.getRows());
-		assertEquals(3, k.offspringBoard.getCols());
-		assertEquals(3, k.offspringBoard.getRows());
-//		//fail("Not yet implemented");
+		assertEquals(3, k.lifeBoard.getCols().intValue());
+		assertEquals(3, k.lifeBoard.getRows().intValue());
+		assertEquals(3, k.offspringBoard.getCols().intValue());
+		assertEquals(3, k.offspringBoard.getRows().intValue());
+		//fail("Not yet implemented");
 	}
 
 }
